# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Medical Booking System API",
        "description": "REST API for the Medical Booking System.",
        "contact": {
            "email": "admin@yourhospital.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/v1/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "List all bookings for the hospital",
                "operationId": "e1b5ea5257f8f23cb7863dd78e44df2c",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter by date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (pending/accepted/etc.)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "doctor_id",
                        "in": "query",
                        "description": "Filter by doctor ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Booking"
                                            }
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking (in-person / admin)",
                "operationId": "46b092e0866a0d48225d1dd0e5d9f22f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_name",
                                    "patient_phone",
                                    "doctor_id",
                                    "booking_date",
                                    "start_time"
                                ],
                                "properties": {
                                    "patient_name": {
                                        "type": "string",
                                        "example": "Vijay Kumar"
                                    },
                                    "patient_phone": {
                                        "type": "string",
                                        "example": "919994780436"
                                    },
                                    "patient_email": {
                                        "type": "string",
                                        "example": "vijay@email.com"
                                    },
                                    "doctor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "booking_date": {
                                        "type": "string",
                                        "example": "2026-03-29"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "example": "10:30"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "cause": {
                                        "type": "string",
                                        "example": "Fever and headache"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "booking_code": {
                                            "type": "string",
                                            "example": "BK-XK92MPLR"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bookings/{id}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get a single booking by ID",
                "operationId": "4ce45cf16c092ff3ac78bd9da2d5db20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bookings/{id}/status": {
            "put": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Update booking status (accept / reject / complete etc.)",
                "operationId": "5fac768bbdbf6cae33dd8e36d490d396",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "accepted",
                                            "rejected",
                                            "cancelled",
                                            "completed",
                                            "no_show"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bookings/{id}/reschedule": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Reschedule a booking to a new date and time",
                "operationId": "0ff3ca86700180486536f39bf550f29e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "new_date",
                                    "new_time"
                                ],
                                "properties": {
                                    "new_date": {
                                        "type": "string",
                                        "example": "2026-04-01"
                                    },
                                    "new_time": {
                                        "type": "string",
                                        "example": "14:30"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Doctor unavailable"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rescheduled successfully"
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bookings/status/{code}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get booking status by booking code (public — no auth needed)",
                "operationId": "6cccf4fb0130c10daba2174f5e52301f",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Booking code e.g. BK-XK92MPLR",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/bookings/cancel/{code}": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Cancel a booking by booking code (public — no auth needed)",
                "description": "Patients can cancel their own booking using their booking code and phone number for verification.",
                "operationId": "2bba44029ce3c6b277d5ad932ee40ba9",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Booking code e.g. BK-XK92MPLR",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "Patient phone number to verify ownership",
                                        "type": "string",
                                        "example": "919994780436"
                                    },
                                    "reason": {
                                        "description": "Optional cancellation reason",
                                        "type": "string",
                                        "example": "Cannot attend"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking cancelled successfully"
                    },
                    "403": {
                        "description": "Phone number does not match booking"
                    },
                    "404": {
                        "description": "Booking not found"
                    },
                    "422": {
                        "description": "Booking cannot be cancelled (already completed or cancelled)"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Booking": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "action_token": {
                        "type": "string",
                        "example": "BK-XK92MPLR"
                    },
                    "patient_name": {
                        "type": "string"
                    },
                    "patient_phone": {
                        "type": "string"
                    },
                    "booking_date": {
                        "type": "string",
                        "example": "2026-03-29"
                    },
                    "start_time": {
                        "type": "string",
                        "example": "10:30:00"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "doctor_name": {
                        "type": "string"
                    },
                    "cause": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Your api_code from the users table. Pass as: Authorization: Bearer {token}",
                "bearerFormat": "API Token",
                "scheme": "bearer"
            }
        }
    },
...
}